created and tested order service, controller, and repository and OrderItem service. order service correctly records user id, date of order, ordered items, address, shipping amount and total price

+++ src/main/java/org/yearup/controllers/ProductsController.java
@@ -8,6 +8,7 @@ import org.springframework.web.server.ResponseStatusException;
 import org.yearup.models.Product;
 import org.yearup.service.ProductService;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -24,8 +25,8 @@ public class ProductsController
     @GetMapping
     @PreAuthorize("permitAll()")
     public List<Product> allProducts(@RequestParam(name="cat", required = false) Integer categoryId,
-                                @RequestParam(name="minPrice", required = false) Double minPrice,
-                                @RequestParam(name="maxPrice", required = false) Double maxPrice,
+                                @RequestParam(name="minPrice", required = false) BigDecimal minPrice,
+                                @RequestParam(name="maxPrice", required = false) BigDecimal maxPrice,
                                 @RequestParam(name="subCategory", required = false) String subCategory) {
         return productService.search(categoryId, minPrice, maxPrice, subCategory);
     }
